requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.3
Brotli==1.0.9
bs4==0.0.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
dash==1.17.0
dash-core-components==1.13.0
dash-html-components==1.1.1
dash-renderer==1.8.3
dash-table==4.11.0
fake-useragent==0.1.11
Flask==1.1.2
Flask-Compress==1.8.0
future==0.18.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.6.2
MarkupSafe==1.1.1
numpy==1.19.4
pandas==1.1.4
parse==1.18.0
plotly==4.13.0
pyee==7.0.4
pyppeteer==0.2.2
pyquery==1.4.3
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
requests-html==0.10.0
retrying==1.3.3
six==1.15.0
soupsieve==2.0.1
scipy==1.5.4
tqdm==4.54.1
urllib3==1.26.2
w3lib==1.22.0
websockets==8.1
Werkzeug==1.0.1
yahoo-fin==0.8.6
